Board class
Board class, implements a board of size X by Y. Contains a print
operation so that we can see the board, as well as a
place_stone(x-location, y-location, boolean is black) function that will
place the stone if possible and update the board, removing all dead
pieces etc. Running time for place_stone is O(n) on the number of moves.