Should work now
I tested it on several test cases, they all seemed to work. This only
test for UNCONDITIONAL life, not for groups that are for all intents and
purposes alive (like the 4-in a row). If someone else wants to test it
on more cases, feel free (or if there was some easy way to generate test
cases: can we modify the gui so that I can just place moves on it and
then have it output the live stones? I mean, the packaging doesn't work
for me still, but I haven't really tried to fix it yet)

If everyone agrees with this then we can merge it.